+++ data/tasks.txt
@@ -1,10 +1,13 @@
-E | 1 | gyat | 2003-09-23T12:40 | 2025-10-12T15:40
+E | 0 | gyat | 2003-09-23T12:40 | 2025-10-12T15:40
 D | 0 | milk | 2025-09-05T23:00
 T | 0 | hellow
 T | 0 | soup
-T | 0 | yap
+T | 1 | yap
 D | 0 | lup | 2024-06-10T14:30
 E | 0 | peep | 2040-12-21T14:50 | 2050-10-23T16:30
 T | 0 | hello
 T | 0 | eat food
 T | 0 | hello
+T | 0 | hello
+D | 0 | hello | 2004-12-31T21:05
+T | 0 | refreshing
