+++ src/main/java/fullmarksbot/FullMarksException.java
@@ -1,0 +1,15 @@
+package fullmarksbot;
+
+/**
+ * Exception thrown for errors specific to FullMarksBot.
+ */
+public class FullMarksException extends Exception {
+    /**
+     * Constructs a FullMarksException with the given message.
+     *
+     * @param message Error message.
+     */
+    public FullMarksException(String message) {
+        super(message);
+    }
+}
