+++ src/main/java/fullmarksbot/Launcher.java
@@ -6,6 +6,11 @@ import javafx.application.Application;
  * A launcher class to workaround classpath issues.
  */
 public class Launcher {
+    /**
+     * Launches the JavaFX application.
+     *
+     * @param args Command-line arguments.
+     */
     public static void main(String[] args) {
         Application.launch(Main.class, args);
     }
