+++ src/main/java/fullmarksbot/Parser.java
@@ -1,5 +1,7 @@
 package fullmarksbot;
 
+import java.util.regex.Pattern;
+
 /**
  * Parses user input into commands and arguments for FullMarksBot.
  */
@@ -20,17 +22,17 @@ public class Parser {
      *
      * @param input User input string containing the task number.
      * @return Zero-based index of the task.
-     * @throws FullMarksBot.FullMarksException If the task number is missing or invalid.
+     * @throws FullMarksException If the task number is missing or invalid.
      */
-    public static int getTaskNumber(String input) throws FullMarksBot.FullMarksException {
+    public static int getTaskNumber(String input) throws FullMarksException {
         String[] parts = input.split(" ");
         if (parts.length < 2) {
-            throw new FullMarksBot.FullMarksException("Please specify a task number.");
+            throw new FullMarksException("Please specify a task number.");
         }
         try {
             return Integer.parseInt(parts[1]) - 1;
         } catch (NumberFormatException e) {
-            throw new FullMarksBot.FullMarksException("Invalid task number. Please enter a number.");
+            throw new FullMarksException("Invalid task number. Please enter a number.");
         }
     }
 
@@ -39,13 +41,25 @@ public class Parser {
      *
      * @param input User input string containing the find command.
      * @return Keyword to search for.
-     * @throws FullMarksBot.FullMarksException If the keyword is missing.
+     * @throws FullMarksException If the keyword is missing.
      */
-    public static String getFindKeyword(String input) throws FullMarksBot.FullMarksException {
+    public static String getFindKeyword(String input) throws FullMarksException {
         String[] parts = input.trim().split(" ", 2);
         if (parts.length < 2 || parts[1].trim().isEmpty()) {
-            throw new FullMarksBot.FullMarksException("Please specify a keyword to find.");
+            throw new FullMarksException("Please specify a keyword to find.");
         }
         return parts[1].trim();
+    }
+
+    /**
+     * Returns true if the input contains the exact word, case-insensitive.
+     *
+     * @param input Input string to search.
+     * @param word Word to search for.
+     * @return True if the word is found as a whole word, false otherwise.
+     */
+    public static boolean containsExactWord(String input, String word) {
+        String pattern = "(?i)\\b" + Pattern.quote(word) + "\\b";
+        return Pattern.compile(pattern).matcher(input).find();
     }
 }
