+++ src/main/java/fullmarksbot/Storage.java
@@ -31,7 +31,7 @@ public class Storage {
         File file = new File(filePath);
         file.getParentFile().mkdirs();
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(file))) {
-            for (FullMarksBot.Task task : taskList.getTasks()) {
+            for (Task task : taskList.getTasks()) {
                 writer.write(task.writeTasks());
                 writer.newLine();
             }
@@ -58,16 +58,16 @@ public class Storage {
                 String[] parts = line.split(" \\| ");
                 String type = parts[0];
                 boolean isDone = parts[1].equals("1");
-                FullMarksBot.Task task = null;
+                Task task = null;
                 switch (type) {
                 case "T":
-                    task = new FullMarksBot.Todo(parts[2]);
+                    task = new Todo(parts[2]);
                     break;
                 case "D":
-                    task = new FullMarksBot.Deadline(parts[2], parts[3]);
+                    task = new Deadline(parts[2], parts[3]);
                     break;
                 case "E":
-                    task = new FullMarksBot.Event(parts[2], parts[3], parts[4]);
+                    task = new Event(parts[2], parts[3], parts[4]);
                     break;
                 }
                 if (task != null && isDone) {
